For multiboot do not rmtree to avoid 'softbrick' boxes
When by accident the partition is still mounted it could remove full
images and/or the STARTUP files that 'softbrick' the boxes. You need to
reinstall images and/or the recovery image. This avoids it in case
something odd happens. Then better get a GSOD instead of 'softbrick' a
box