FlashImage: Further simplification
And solve issue when the file contents needs to be build.

I did this because I did not like having mode 12 as slot values > 12.
Now I have a tupple where the first parameter defines the slot and the
second the mode (when the box has no mode the default is set to mode 1)