No truncate event text for UTF8
Author of the original code PiDron(gisclub.tv)

-This problem is related to the fact that when saving information from
SHORT_EVENT_DESCRIPTOR to the epg.dat cache file, it is transcoded to
UTF-8. And since the information is initially transmitted in ISO 8859-5
encoding, in which the Cyrillic character is encoded in one byte - the
standard field SHORT_EVENT_DESCRIPTOR of 255 - 6 bytes can contain text
in 249 characters, in UTF-8 Cyrillic characters are encoded in two bytes
and, accordingly, in the field SHORT_EVENT_DESCRIPTOR can fit
approximately 125 Cyrillic characters, so that everything that is more
than 249 bytes is truncated.