[HdmiCec] avoid echoing 'System Standby' command back to TV
When System Standby (36) is received, e2 switches to standby.
As a result of going into standby, e2 would transmit another System
Standby command back to the TV.

Avoid that by setting a 'handlingStandbyFromTV' boolean while handling a
System Standby command and not transmitting 'System Standby' while the
variable is set.