filepush: resolve deadlock
The extra poll() added for Broadcom transcoding would consume the
SIGUSR1 in the processing thread when sent by the parent, it would
get ignored and the read() would be entered and it would never return.

Resolved this by using ppoll() and checking the return value of it,
so the SIGUSR1 doesn't get ignored.

There is no need to add more tests on the stop condition, just don't
ignore the signal is enough to solve it.

Thanks to SimonC for spotting the condition and helping analyse it.